Implement password functionality, revert navbar changes, other fixes
